Update homepage copy
I'm giving our copy a logical flow. First we're telling them in bold letters that this is the fastest way for them to get coding. Our most popular pages and requests for info are about code schools, scholarships, and jobs. So it makes sense for our copy to say we'll provide them that and more. The copy changes will also help with SEO.