Update copy on mentorship page
Making it more concise so the information is more digestible. For example, we already say folks can choose any programming language/stack of their choice so just clutters the page to include an unordered list of some of their options. Also, it makes more sense to simply refer to the program as our Software Mentorship Program instead of the Software Mentor Protege program since we already call it the software mentorship program on the homepage, in the nav menu and in the copy.