Amended meta tag according to #540 (#543)
* Amended meta tag according to #540

Substituted a few words and rearranged to convey the same message in a more compact format

* Added "vets" & "military vets" to keywords

Maybe those terms aren't the most preferred for some, many of my family and friends who are veterans do refer to themselves as vets.  Adding this to both the meta tag and including in keywords should broaden the scope for SEO contributing to the overall goal of the project.